Publish the NuGet package also on GitHub
Now that I tested the GitHub Package Registry there's version 1.0.0-alpha.0.97 which can not be deleted and is displayed on the main page of the repository. I can see how this could cause confusion if it stays at an alpha version while the final version is published on nuget.org. So I'll publish releases on GitHub too in order to have the package display a final version on the GitHub project page. Nobody will use it since it requires adding a custom NuGet feed and requires authentication but at least it will generate confusion about what is the currently published version.