Use short RepositoryUrl (without .git suffix)
Locally, the .git/config file was configured with a scp-like url:
```
[remote "origin"]
	url = git@github.com:0xced/serilog-formatting-log4net.git
```

This would produce a RepositoryUrl ending in .git which does not match the one on the continuous integration server which is configured with https.

Changing the `origin` remote to https://github.com/0xced/serilog-formatting-log4net ensures that the RepositoryUrl is the same when running locally and on the continuous integration server.