Add net6.0 target framework
It will be beneficial for [string interpolation][1] which is used extensively through the project. See [Are there benefits in producing a .NET 6.0 version of a .NET Standard 2.0 library?][1] on Stack Overflow.

Also, run mutation tests _after_ creating the NuGet package to ensure that Serilog.Formatting.Log4Net.dll is not altered by Styker.NET.

[1]: https://devblogs.microsoft.com/dotnet/performance-improvements-in-net-6/#arrays-strings-spans
[2]: https://stackoverflow.com/questions/70778797/are-there-benefits-in-producing-a-net-6-0-version-of-a-net-standard-2-0-librar/72266562#72266562