Use Constructor / Dispose instead of [Before(Test)] / [After(Test)]
Notes below are for TUnit version 0.25.21. As of version 1.0.78, Constructor / Dispose works the same as Before / After attributes.

Note that this fails the two test with the [NotInParallel] attribute, i.e. FilterPropertyThrowing and ExceptionFormatterThrowing.

The [Test Set Ups][1] and [Test Clean Ups][2] documentation mentions respectively:

> Most setup for a test can be performed in the constructor (think setting up mocks, assigning fields.)

and

> TUnit supports having your test class implement `IDisposable` or `IAsyncDisposable`. These will be called after your test has finished executing.

So I'm not sure what's the difference between Custructor/Dispose and [Before(Test)] / [After(Test)] attributes.

[1]: https://thomhurst.github.io/TUnit/docs/tutorial-extras/setup
[2]: https://thomhurst.github.io/TUnit/docs/tutorial-extras/cleanup